==> 01_2_train_item_description_word2vec.ipynb <==
## Train Custom Word2Vec Model on Product Descriptions

Word vectors trained to represent the use in product description domain - results are averaged to represent item as average of word vectors in product description

import pandas as pd
import numpy as np

import gensim
from gensim.models import FastText
from gensim.models import Word2Vec
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
import nltk
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *

nltk.download('wordnet')

data_all = pd.read_pickle("data/processed_data/articles_info.pkl")

data_all.dtypes

data = data_all.text_clean.tolist()
len(data)

Create Stopwords to delete from descriptions

my_stop_words = STOPWORDS.union(set(['size', 'color', 'material', 'product', 'dimension', 'length', 'package', 'brand', 
                     'pack', 'width', 'piece', 'height', 'quality', 'group', 'high', 'model', 'article', 'assort',
                     'price', 'weight', 'colour', 'products', 'type', 'design', 'diameter'])) 
# retrieved from previous project

my_stop_words

Define and test process: 
- remove stopwords
- remove words less than 4 characters
- lemmatize words

def preprocess(text):
    result = []
    for token in gensim.utils.simple_preprocess(text):
        if token not in my_stop_words and len(token) > 3: # standard stopwords and remove shorter 3 
            result.append(WordNetLemmatizer().lemmatize(token, pos='v'))
    return result

preprocess(data[500])

#preprocessed_data = data_all['text_clean'].map(preprocess)
preprocessed_data = [preprocess(str(data[i])) for i in range(len(data))]

Train word2vec model for all words that are in the corpus at least 3 times

model = Word2Vec(preprocessed_data, size=150, window=5, min_count=3, workers=4) 
model.save("data/models/word2vec/word2vec.model")
# killed on my computer.. try again on Google Cloud

model.wv.most_similar('christmas')

model.wv['cup']

word_vectors = model.wv

Function to calculate the average vector across all words in the description

def get_mean_vector(word2vec_model, words):
    # remove out-of-vocabulary words
    words = [word for word in words if word in word_vectors.vocab]
    if len(words) >= 1:
        return np.mean(word2vec_model[words], axis=0)
    else:
        return []

get_mean_vector(model, preprocessed_data[1])

item_vec = [get_mean_vector(model, preprocessed_data[i]) for i in range(len(preprocessed_data))]

len(item_vec)

np.save("data/processed_data/item_text_vec.npy", item_vec)

## Test Performance

from sklearn.metrics.pairwise import cosine_similarity

# testing it against the best shown from word2vec it seems to work better ! some of the weird ones there dont get high scores
# here !!

sim = cosine_similarity(item_vec[99999].reshape(1,-1), item_vec[569553].reshape(1,-1))
sim

## Save array in item data

df = pd.DataFrame({"text_vec": item_vec})

df.head()

item_text_vec = pd.merge(data_all, df, left_index = True, right_index= True)

len(item_text_vec)

item_text_vec.head()

item_text_vec.to_pickle('data/processed_data/articles_info.pkl')

np.array(item_text_vec.text_vec[0:10])

test = item_text_vec.text_vec[0:10]

sim = cosine_similarity(item_text_vec.text_vec[99999].reshape(1,-1), test.tolist())
sim
