==> 01_3_create_target_data.ipynb <==
# Create Target Table

- join clicks and pick table in order to receive customer "rating" feedback concerning a product
- all items that are picked by the user = added to the shopping cart, are rated positive, the others negative
- divide data in training and testing data

import pandas as pd 
import numpy as np 
import pickle
import feather

# Clicks

A click is measured when a user clicks on a product on a listing page - it leads the user to the product detail page

clicks = feather.read_dataframe("data/raw_data/artikelklicks_current.feather")

clicks.head()

len(clicks) 

max(clicks.datum)

len(clicks.userID.unique())

get unique combinations of userID and artikelID

len(clicks.groupby(['userID','artikelID']).size().reset_index())


get unique combinations of userID and anbieter_artikelnummer - a little more general than artikelID

len(clicks.groupby(['userID','anbieter_artikelnummer']).size().reset_index())


clicks.datum = pd.to_datetime(clicks.datum)

Only include each item and user combination once during training

clicks_unique = clicks.groupby(['userID','anbieter_artikelnummer'])['datum', 'anbieterID'].min().reset_index()

clicks_unique.head()

# Picks

a pick is the target variable of this model - it signals when the user adds a product to the shopping cart

picks = pd.read_csv("raw_data/artikelpicks.csv", sep = ";", encoding = "unicode_escape")
picks.head()

len(picks)

max(picks.datum)

# number of unique user and item combinations!
len(picks.groupby(['userID','anbieter_artikelnummer']).size().reset_index())


picks.datum = pd.to_datetime(picks.datum, utc = True)


picks.head()

len(picks.groupby(['userID','anbieter_artikelnummer', 'date'])['datum'].min().reset_index())

picks summarized on user and anbieter_artikelnummer - ready for outer-join with clicks

picks_unique = picks.groupby(['userID','anbieter_artikelnummer'])['datum', 'anbieterID'].min().reset_index()
picks_unique['pick'] = 1

print(len(picks_unique))
picks_unique.head()

# Merge clicks and picks as outer join on userID, anbieter_artikelnummer

len(clicks_unique)

len(picks_unique)

target = pd.merge(clicks_unique, picks_unique, how = "outer", on = ["userID", "anbieter_artikelnummer", "anbieterID"], suffixes = ("_click", "_pick")) 

(target.sample(100))

target.pick.sum() # works !

Fill NAs: 
- all picks that are not 1 need to be 0
- all rows that do not have a click date need to fill the click that with pick date (some items were picked without looking on the item detail page)


target.fillna(value = {'pick': 0}, inplace = True)
target.loc[pd.isnull(target.datum_click) == True, 'datum_click'] = target.datum_pick
#target.loc[target.datum_click == 0, 'datum_click'] = target.datum_pick
target.sample(100)

target.to_pickle("data/processed_data/target_data_all.pkl")

# Analyse Target Data

How many differnt articles, users, how often do they appear

target = pd.read_pickle("data/processed_data/target_data_all.pkl")

len(target)

target_top = target.head(1000)

# Users for training
target['userID'].nunique()


# Articles for training
target['anbieter_artikelnummer'].nunique()


# number of positive feedback
target.pick.sum()

# pick-ratio
target.pick.sum()/len(target)

## Users

find ouliers and filter training data

target_summary_user = target.groupby(['userID'])['datum_pick', 'pick'].count().reset_index()

target_summary_user.columns = ['userID', 'picks', 'clicks']
target_summary_user.describe()

# users with only one click
(sum(target_summary_user.clicks ==1))

14% of users only made one click

(sum(target_summary_user.clicks ==1))/len(target_summary_user)

target_summary_user.sort_values(by=['picks'], ascending = False).head(10)

Many of these must be hackers that are scraping our website... no one clicks this much an never picks

hackers = target_summary_user.sort_values(by=['clicks'], ascending = False).head(50)
hackers

should filter out - only top 9 is suspicious

### Clean and Filter Users

# look at obvious hackers = thousands of clicks - no picks... not usuall behavior 
hackers_only = hackers[hackers['picks']==0]

userIDs_filter = hackers_only.userID.tolist()

target_filtered = target[~target.userID.isin(userIDs_filter)]

len(target_filtered) #13289785 before

## Articles

target_summary_article = target.groupby(['anbieter_artikelnummer'])['datum_pick', 'pick'].count().reset_index()

target_summary_article.columns = ['anbieter_artikelnummer', 'picks', 'clicks']
target_summary_article.describe()

(sum(target_summary_article.clicks ==1))

(sum(target_summary_article.clicks ==1))/len(target_summary_article)

target_summary_article.sort_values(by=['clicks'], ascending = False).head(10)

useless_articles = ['']

### Clean and filter articles

target_filtered = target_filtered[~target_filtered.anbieter_artikelnummer.isin(useless_articles)]

len(target_filtered) # 12573323

# Filtered Dataset - final for now 27.9.19

target_filtered.to_pickle("data/processed_data/target_data_filtered.pkl")

# Users & Items

extract userIDs that need embeddings

target_filtered['userID'].nunique()

users = target_filtered['userID'].unique()

target_filtered['anbieter_artikelnummer'].nunique()

articles = target_filtered['anbieter_artikelnummer'].unique()

with open('data/processed_data/users_list.pkl', 'wb') as f:
    pickle.dump(users, f)
    f.close()

with open('data/processed_data/articles_list.pkl', 'wb') as f:
    pickle.dump(articles, f)
    f.close()

# Train and Test Split

Create train and test split based on time - first 90% of data is training, the last 10% is testing data

target_filtered = target_filtered.sort_values(by='datum_click')

test_size = len(target_filtered)*0.1

test_size

target_test = target_filtered[-1251600:]

len(target_test)

target_train = target_filtered[:-1251600]

len(target_train)

len(target_train)+len(target_test) == len(target_filtered)

target_test.sort_values(by='datum_click')

target_train.sort_values(by='datum_click', ascending = False)

target_train.to_pickle("data/training_data/target_train.pkl")

target_test.to_pickle("data/testing_data/target_test.pkl")
