==> 01_4_create_features_log_training.ipynb <==
# Create Training Data

Calculate context features for the training data
- days online
- month of the year

# Items

## Days online

how many days has the item been available on zentrada at the moment the user clicks the item

import pandas as pd
import numpy as np
import pickle 
from datetime import timedelta

item_data = pd.read_pickle('data/training_data/articles_info_final.pkl')

log = pd.read_pickle('data/training_data/target_train.pkl')

item_data.head()

item_data = item_data[['anbieter_artikelnummer', 'erstanlageDatum']]

df = pd.merge(log, item_data, how = 'left' )

df.head()

test = df.sample(20).reset_index()

test.dtypes

df.datum_click[1] - df.erstanlageDatum[1]

Calculate the time online

df['online'] = (df.datum_click - df.erstanlageDatum)

df.head()

Convert it in days

df['days_online'] = [df.online[i].days for i in range(len(df))]

df.drop('online', axis = 1, inplace = True)
df.head()

df.to_pickle('data/training_data/target_train_process.pkl')

## Month of the year

Calculate the month of the year at time of item click - show seasonality

df = pd.read_pickle('data/training_data/target_train_process.pkl')

df.head()

df.drop(['anbieterID', 'datum_pick', 'erstanlageDatum'], axis = 1, inplace = True)

df.head()

extract month

df['month'] = [df.datum_click[i].month for i in range(len(df))]

df.head()

df.to_pickle('data/training_data/target_train_done.pkl')

# User

Add the user's marketplace to the log data

df.head()

user = pd.read_csv("data/raw_data/userstammdaten.csv", sep = ";", encoding = 'ISO-8859-15')

user.head()

user = user[['userID', 'erstRegMarktplatz']]

df = pd.merge(df, user, how = "left", on = "userID")

df.head()

len(df)

sum(pd.isnull(df['erstRegMarktplatz'] == True))

len(df[pd.isnull(df['erstRegMarktplatz']) == False])

df.to_pickle('data/training_data/target_train_done.pkl')
