==> 02_0_statistics_training_data.ipynb <==
# Statistics Final Training Data

These values show the real distribution, running this sheet will show the distributions of the sample data set

import pandas as pd
import sqlalchemy 
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy import Column, Integer, String, DateTime, Float
from sqlalchemy import create_engine, MetaData, Table, inspect
import psycopg2
from sqlalchemy.dialects import postgresql
from matplotlib import pyplot as plt 
import sqlite3

#DATABASE_URI = ###
#engine = create_engine(DATABASE_URI)

engine = sqlite3.connect('data/db.db')

## Density

users = pd.read_sql('''SELECT DISTINCT t.userid
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL''', engine)

items = pd.read_sql('''SELECT DISTINCT t.anbieter_artikelnummer
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL''', engine)

clicks = pd.read_sql('''SELECT count(*)
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL''', engine)

len(users)

len(items)

clicks

## Clicks per User

users = pd.read_sql('''SELECT t.userid, count(*)
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL GROUP BY t.userid''', engine)

users.columns = ['userid', 'count']

users.sort_values('count', ascending = False).head(10)

users['count'].describe()

Percentage of users with more than 200 clicks

len(users[users['count'] >=200])/len(users)

## Clicks per Item

items = pd.read_sql('''SELECT t.anbieter_artikelnummer, count(*)
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL GROUP BY t.anbieter_artikelnummer''', engine)

items.columns = ['anbieter_artikelnummer', 'count']

items.sort_values('count', ascending = False).head(10)

items['count'].describe()

## Pick-Ratio

picks = pd.read_sql('''SELECT sum(t.pick)
            FROM    target_training_enc as t 
                    LEFT JOIN training_user_filter as f
                        ON t.userid = f.userid
            WHERE   f.userid IS NULL''', engine)

picks

Pick-Ratio

picks.values[0][0]/clicks.values[0][0]

pd.read_sql("SELECT * from user_enc where userid = '1698332'", engine)

import pickle

imputer_stueck = pickle.load( open( "data/models/preprocessing/imputer_stueck.pkl", "rb" ) )
imputer_preis = pickle.load( open( "data/models/preprocessing/imputer_preis.pkl", "rb" ) )

imputer_stueck.transform([[None]])

imputer_preis.transform([[None]])
