==> 02_1_transform_item_features.ipynb <==
# Transform Item Features for Model

import pandas as pd
import numpy as np
import pickle 

## Postgres: Create Table item_enc 

import sqlalchemy 
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy import Column, Integer, String, DateTime, Float
import psycopg2
from sqlalchemy.dialects import postgresql

DATABASE_URI = ###
from sqlalchemy import create_engine, MetaData, Table, inspect
engine = create_engine(DATABASE_URI)

# Connect to already existing db table
meta = MetaData(engine)
item_vec = Table('item_enc', meta, autoload=True)

# if need to delete table
#item_enc.drop(engine)

Create table

meta = MetaData(engine)
item_enc = Table('item_enc', meta,
                 Column('anbieter_artikelnummer', String, primary_key=True),
                 Column('anbieterid_enc', Integer),
                 Column('anbietermarktplatz_enc', Integer),
                 Column('warengruppe_enc', Integer),
                 Column('text_vec', postgresql.ARRAY(Float)),
                 Column('preis_std', Float),
                 Column('minve_std', Float))
item_enc.create()

## Sqlite3: Create Table item_enc 

engine = sqlite3.connect('data/db.db')

c = engine.cursor()

c.execute('''CREATE TABLE item_enc
             ([anbieter_artikelnummer] text PRIMARY_KEY, [anbieterid_enc] integer, [anbietermarktplatz_enc] integer,
             [warengruppe_enc] integer, [text_vec] blob, [preis_std] real, [minve_std] real, [preis_log_std] real, [minve_log_std] real, [erstanlagedatum] datetime  )''')

## Import Item Data

item_data = pd.read_pickle('data/training_data/articles_info_final.pkl')

item_data[-50:]

len(item_data)

sum(pd.isna(item_data.erstanlageDatum))

#item_data = item_data[:10000]

## Train and Save Imputer, Encoder, and StandardScaler

from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
#from sklearn.preprocessing import OneHotEncoder 
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import LabelEncoder

### Categorical Data

anbieterID, anbietermarktplatz, warengruppe

# anbieterID
label_enc_id = LabelEncoder()

#Calling methods on our OneHotEncoder object
label_enc_id.fit( item_data.anbieterID.values) 
pickle.dump(label_enc_id, open("data/models/preprocessing/label_anbieterID.pkl", "wb"))


# anbietermarktplatz
label_enc_mkt = LabelEncoder()

#Calling methods on our OneHotEncoder object
label_enc_mkt.fit( item_data.anbietermarktplatz.values) 
pickle.dump(label_enc_mkt, open("data/models/preprocessing/label_mkt.pkl", "wb"))


# warengruppe
item_data.loc[:,'warengruppe'] = item_data.warengruppe.fillna("NN")
label_enc_wg = LabelEncoder()

#Calling methods on our OneHotEncoder object
label_enc_wg.fit( item_data.warengruppe.values) 
pickle.dump(label_enc_wg, open("data/models/preprocessing/label_warengruppe.pkl", "wb"))


### Continouus Data

stueck_pro_ve, preis_euro - realized that i needed to delete more outliers, trained new transformers in 02_2

# stueck_pro_ve
imputer_stueck = SimpleImputer (strategy = "mean")

imputer_stueck.fit(item_data[['stueck_pro_ve']])#.values.reshape(-1.1))
pickle.dump(imputer_stueck, open("data/models/preprocessing/imputer_stueck.pkl", "wb"))
transformed_stueck = imputer_stueck.transform(item_data[['stueck_pro_ve']])

std_stueck = StandardScaler()

std_stueck.fit(transformed_stueck)
pickle.dump(std_stueck, open("data/models/preprocessing/scaler_stueck.pkl", "wb"))
#transformed_cont = std.transform(transformed_cont)

# preis_euro
imputer_preis = SimpleImputer (strategy = "mean")

imputer_preis.fit(item_data[['preis_euro']])#.values.reshape(-1.1))
pickle.dump(imputer_preis, open("data/models/preprocessing/imputer_preis.pkl", "wb"))
transformed_preis = imputer_preis.transform(item_data[['preis_euro']])

std_preis = StandardScaler()

std_preis.fit(transformed_preis)
pickle.dump(std_preis, open("data/models/preprocessing/scaler_preis.pkl", "wb"))
#transformed_cont = std.transform(transformed_cont)

## Pipeline

test = item_data[0:1000].reset

test

# load all transformers
# cat
label_enc_id = pickle.load( open( "data/models/preprocessing/label_anbieterID.pkl", "rb" ) )
label_enc_mkt = pickle.load( open( "data/models/preprocessing/label_mkt.pkl", "rb" ) )
label_enc_wg = pickle.load( open( "data/models/preprocessing/label_warengruppe.pkl", "rb" ) )
# cont
imputer_stueck = pickle.load( open( "data/models/preprocessing/imputer_stueck.pkl", "rb" ) )
imputer_preis = pickle.load( open( "data/models/preprocessing/imputer_preis.pkl", "rb" ) )
std_stueck = pickle.load( open( "data/models/preprocessing/scaler_stueck.pkl", "rb" ) )
std_preis = pickle.load( open( "data/models/preprocessing/scaler_preis.pkl", "rb" ) )


def transform_data(df):
    # fill empty warengruppe
    #warengruppe = df.warengruppe.fillna("NN")
    df1 = pd.DataFrame({'anbieter_artikelnummer': df.anbieter_artikelnummer})
    # impute and scale preis and stueck_ve
    df1["minVE_std"] = imputer_stueck.transform(df[['stueck_pro_ve']])
    df1["preis_std"] = imputer_preis.transform(df[['preis_euro']])
    df1["minVE_std"] = std_stueck.transform(df1[['minVE_std']])
    df1["preis_std"] = std_preis.transform(df1[['preis_std']])
    # Create one-hot-encodings anbieterID, anbietermarktplatz, warengruppe
    anbieterID = (label_enc_id.transform(df['anbieterID']) + 1).tolist()
    anbietermarktplatz = (label_enc_mkt.transform(df['anbietermarktplatz']) + 1).tolist()
    warengruppe = (label_enc_wg.transform(df.warengruppe.fillna("NN").values) + 1).tolist()
    text_vec_list = []
    # Change text_vec to list
    for n in range(len(df.text_vec)):
        try:
            ls = df.text_vec[n].tolist()
        except: 
            ls = [0] * 150
        finally:
            text_vec_list.append(ls)
    # combine data to second df
    df2 = pd.DataFrame({"anbieter_artikelnummer": df.anbieter_artikelnummer,
                        'text_vec': text_vec_list, 
                        "anbieterID_enc" : anbieterID, 
                        "anbietermarktplatz_enc": anbietermarktplatz, 
                        "warengruppe_enc": warengruppe})
    df = pd.merge(df1, df2, how = "left", on = "anbieter_artikelnummer")
    return(df)

transform_data(item_data[-50:].reset_index())

## Transform and Save Data in DB

use batches in order to make manageable

batches_start = list(range(0, len(item_data), 50000))
batches_end = batches_start[1:]
batches_end.append(len(item_data))

for i,v in zip(batches_start, batches_end):
    transformed = transform_data(item_data[i:v].reset_index())
    # save data
    transformed.to_sql('item_enc', engine, index = False, if_exists = 'append') # later need to check first if already in DB

transformed

Test if extraction works - yes
