==> 06_1_word2vec_visualization.ipynb <==
## Create Vectors for Visualization in Embedding Projector

resulting files are in google drive folder visualization_embeddings. 
- text_vecs.tsv
- text_labels_tsv

Upload to https://fontjoy.com/projector/ to visualize the embeddings

import csv
import sqlalchemy as db
from sqlalchemy import create_engine
import psycopg2
import pandas as pd
import numpy as np
import sqlite3

# DATABASE_URI = ###
# engine = create_engine(DATABASE_URI)
conn = sqlite3.connect('data/db.db')

oberwarengruppen = pd.read_csv('data/raw_data/warengruppe.csv', sep = ";")

oberwarengruppen.sample()

### Load example items

items = pd.read_sql("SELECT e.anbieter_artikelnummer, e.text_vec, r.warengruppe, r.text_clean from item_enc as e left join item_features_raw as r ON e.anbieter_artikelnummer = r.anbieter_artikelnummer where text_clean NOT IN ('None') limit 10000", engine)

items = items.merge(oberwarengruppen, left_on = "warengruppe", right_on = "warenuntergruppe",  how = 'left')

items['description'] = items['text_clean'].str[:50]

items.head()

list_text = []
for i in range(len(items)):
    list_text.append(items.text_vec[i])
item_text = np.array(list_text, ndmin = 2)

save file for text vectors

np.savetxt('data/models/text_vecs.tsv', item_text, delimiter = "\t" )

save file for labels

subset = items[['anbieter_artikelnummer', 'warengruppe', 'warenobergruppe', 'description']]

meta = [tuple(x) for x in subset.values]

def save_meta():
    label_file = "data/models/text_labels.tsv"
    with open(label_file, 'w', encoding='utf8', newline='') as tsv_file:
        tsv_writer = csv.writer(tsv_file, delimiter='\t', lineterminator='\n')
        tsv_writer.writerow(["anbieter_artikelnummer", "warengruppe", "warenobergruppe", "text"])
        for nummer, unter, ober, text in meta:
            tsv_writer.writerow([nummer, unter, ober, text])

save_meta()

items.groupby('warenobergruppe').anbieter_artikelnummer.count()

## Users

user = pd.read_sql("SELECT userid, text_vec_user from user_enc where preis_std_user !=0 limit 10000", engine)

user

list_text_user = []
for i in range(len(user)):
    list_text_user.append(user.text_vec_user[i])
list_text_user = np.array(list_text_user, ndmin = 2)

np.savetxt('data/models/user_text_vecs.tsv', item_text, delimiter = "\t" )
